% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{eml_twochoiceRL}
\alias{eml_twochoiceRL}
\title{Expectation-Maximization with Laplace approximation (EML) of behavior in a two-choice decision task}
\usage{
eml_twochoiceRL(
  data = NULL,
  param = list(init_x1, init_x2),
  prior_mean = c(m1, m2),
  prior_sd = c(s1, s2),
  fn = "eml.objectiveFunction",
  opt = "TRM",
  radius = c(tr_rad[1], tr_rad[2]),
  nRes = nRes,
  iter = iter,
  eml_data = eml_data,
  progress_bar = progress_bar
)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{param}{randomly generated initial parameters.}

\item{prior_mean}{mean priors for x1 and x2, respectively. Defaults to mean 0 for both parameters.}

\item{prior_sd}{standard deviation priors for x1 and x2, respectively. Defaults to 5 for both parameters.}

\item{fn}{objective function being minimized}

\item{opt}{optimization algorithm used. Defaults to trust-region method (trm)}

\item{radius}{starting and maximum allowed trust region radius. Defaults to 1 and 5.}

\item{nRes}{number of restarts. Defaults to 5.}

\item{iter}{iteration of algorithm}

\item{eml_data}{data to be passed for succeeding iterations}

\item{progress_bar}{track completion time of estimation. Defaults to TRUE.}
}
\value{
A list containing the sum log-likelihood, eml estimates of parameter 1 per subject,
        eml estimates of parameter 2 per subject, laplace values of parameter 1 per subject,
        laplace values of parameter 2 per subject.
}
\description{
This function runs the EML technique to recover behavior in a two-choice decision task.
}
\keyword{internal}
