% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{map_twochoiceRL}
\alias{map_twochoiceRL}
\title{Maximum A Posteriori (MAP) of behavior in a two-choice decision task}
\usage{
map_twochoiceRL(
  data = NULL,
  param = list(init.x1, init.x2),
  prior.mean = c(0, 0),
  prior.sd = c(5, 5),
  fn = "map.objectiveFunction",
  opt = "TRM",
  nRes = 5
)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{param}{randomly generated initial parameters.}

\item{prior.mean}{mean priors for x1 and x2, respectively. Defaults to mean 0 for both parameters.}

\item{prior.sd}{standard deviation priors for x1 and x2, respectively. Defaults to 5 for both parameters.}

\item{fn}{objective function being minimized}

\item{opt}{optimization algorithm used. Defaults to trust-region method (trm)}

\item{nRes}{number of restarts. Defaults to 5.}
}
\value{
A list containing the sum log-likelihood, map estimates of parameter 1 per subject,
        map estimates of parameter 2 per subject, laplace values of parameter 1 per subject,
        laplace values of parameter 2 per subject.
}
\description{
This function runs the MAP technique to recover behavior in a two-choice decision task.
}
\examples{
map_twochoiceRL()
}
\keyword{internal}
