% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{mle_twochoiceRL}
\alias{mle_twochoiceRL}
\title{Maximum-likelihood estimation (MLE) of behavior in a two-choice decision task}
\usage{
mle_twochoiceRL(
  data = NULL,
  param = list(init_x1, init_x2),
  fn = "mle.objectiveFunction",
  opt = "TRM",
  radius = c(tr_rad[1], tr_rad[2]),
  nRes = nRes,
  progress_bar = progress_bar
)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{param}{randomly generated initial parameters}

\item{fn}{objective function being minimized}

\item{opt}{optimization algorithm used. Defaults to trust-region method (trm)}

\item{radius}{starting and maximum allowed trust region radius. Defaults to 1 and 5.}

\item{nRes}{number of restarts. Defaults to 5}

\item{progress_bar}{tracks completion time of function. Defaults to TRUE.}
}
\value{
A list containing the sum log-likelihood, mle estimates of parameter 1 per subject,
        mle estimates of parameter 2 per subject, laplace values of parameter 1 per subject,
        laplace values of parameter 2 per subject.
}
\description{
This function runs the MLE technique to recover behavior in a two-choice decision task.
}
\keyword{internal}
