% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twochoiceRL.R
\name{plot_twochoiceRL}
\alias{plot_twochoiceRL}
\title{Visualize behavior in a two-choice decision task}
\usage{
plot_twochoiceRL(
  data = NULL,
  subj = 100,
  colors = c("orange", "purple"),
  plot_type = "static"
)
}
\arguments{
\item{data}{simulated task data}

\item{subj}{subject data to plot. Defaults to subject 100.}

\item{colors}{color to represent the two choices - choice A or choice B. Defaults to orange for choice A and purple for choice B.}

\item{plot_type}{produce either static or animated version of the plot}
}
\value{
A plot of individual \code{subj} behavior data.
}
\description{
This function plots the simulated two-choice task data.
}
\examples{

# Save simulated task data to a variable, say, data
data <- simulate_twochoiceRL(trials_unique = FALSE)

# Plot the behavior of individual 34
plot_twochoiceRL(data = data, subj = 34, colors = c("#749dae","#5c1a33"))
}
