% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_twochoiceRL.R
\name{softmax}
\alias{softmax}
\title{Softmax decision model}
\usage{
softmax(x)
}
\arguments{
\item{x}{expected value}
}
\value{
choice probability value
}
\description{
This function converts expected value to choice probability.
}
\keyword{internal}
