% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twochoiceRL.R
\docType{package}
\name{twochoiceRL}
\alias{twochoiceRL}
\title{twochoiceRL: Modeling Two-Choice Decision Behavior}
\description{
The twochoiceRL package provides three important functions to
\code{\link[twochoiceRL:simulate_twochoiceRL]{simulate}},
\code{\link[twochoiceRL:plot_twochoiceRL]{visualize}} and
\code{\link[twochoiceRL:estimate_twochoiceRL]{estimate}} two-choice decision
behavior.
}
\seealso{
Resources: \itemize{ \item{\href{https://www.tqmp.org/}{Suthaharan,
  P., Corlett, P.R., & Ang, Y.S. (2021). Computational modeling of behavioral
  tasks: An illustration on a classic reinforcement learning paradigm}}
  \item{\url{https://github.com/psuthaharan/twochoiceRL}} }
}
\author{
\strong{Maintainer}: Praveen Suthaharan
  \email{praveen.suthaharan@yale.edu}

  Authors: \itemize{ \item{Praveen Suthaharan} \item{Philip R. Corlett}
  \item{Yuen-Siang Ang} }
}
