% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{eml_twochoiceRL}
\alias{eml_twochoiceRL}
\title{Expectation-Maximization with Laplace approximation (EML) of behavior in a two-choice decision task}
\usage{
eml_twochoiceRL(
  data = NULL,
  param = list(init.x1, init.x2),
  prior.mean = c(0, 0),
  prior.sd = c(5, 5),
  fn = "eml.objectiveFunction",
  opt = "TRM",
  nRes = 5,
  iter = iter,
  eml_data = eml_data
)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{param}{randomly generated initial parameters.}

\item{prior.mean}{mean priors for x1 and x2, respectively. Defaults to mean 0 for both parameters.}

\item{prior.sd}{standard deviation priors for x1 and x2, respectively. Defaults to 5 for both parameters.}

\item{fn}{objective function being minimized}

\item{opt}{optimization algorithm used. Defaults to trust-region method (trm)}

\item{nRes}{number of restarts. Defaults to 5.}

\item{iter}{iteration of algorithm}

\item{eml_data}{data to be passed for succeeding iterations}
}
\value{
A list containing the sum log-likelihood, eml estimates of parameter 1 per subject,
        eml estimates of parameter 2 per subject, laplace values of parameter 1 per subject,
        laplace values of parameter 2 per subject.
}
\description{
This function runs the EML technique to recover behavior in a two-choice decision task.
}
\examples{
eml_twochoiceRL()
}
\keyword{internal}
