% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{estimate_twochoiceRL}
\alias{estimate_twochoiceRL}
\title{Estimate behavior in a two-choice decision task}
\usage{
estimate_twochoiceRL(data = NULL, method = "mle", plot = TRUE)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{method}{parameter estimation technique used; either maximum-likelihood estimation (mle),
maximum a posteriori (map) or expectation-maximization with laplace approximation (eml).
Defaults to mle.}

\item{plot}{visualize estimation performance between true parameters vs estimated parameters.
Defaults to TRUE.}
}
\value{
A plot illustrating the correlation between the true parameters and estimated parameters, and a list
        containing a dataframe of the true and estimated parameter values, the correlation value of
        parameter 1, and the correlation value of parameter 2.
}
\description{
This function performs parameter estimation to recover behavior in a two-choice decision task
}
\examples{

# Save simulated task data to a variable, say, data
data <- simulate_twochoiceRL(trials.unique = TRUE)

# Recover behavioral parameters using maximum-likelihood estimation (MLE)
estimate_twochoiceRL(data = data, method = "mle", plot=TRUE)
}
