% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_twochoiceRL.R
\name{mle_twochoiceRL}
\alias{mle_twochoiceRL}
\title{Maximum-likelihood estimation (MLE) of behavior in a two-choice decision task}
\usage{
mle_twochoiceRL(
  data = NULL,
  param = list(init.x1, init.x2),
  fn = "mle.objectiveFunction",
  opt = "TRM",
  nRes = 5
)
}
\arguments{
\item{data}{simulated two-choice task data}

\item{param}{randomly generated initial parameters.}

\item{fn}{objective function being minimized}

\item{opt}{optimization algorithm used. Defaults to trust-region method (trm)}

\item{nRes}{number of restarts. Defaults to 5.}
}
\value{
A list containing the sum log-likelihood, mle estimates of parameter 1 per subject,
        mle estimates of parameter 2 per subject, laplace values of parameter 1 per subject,
        laplace values of parameter 2 per subject.
}
\description{
This function runs the MLE technique to recover behavior in a two-choice decision task.
}
\examples{
mle_twochoiceRL()
}
\keyword{internal}
