% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twochoiceRL.R
\name{plot_twochoiceRL}
\alias{plot_twochoiceRL}
\title{Visualize behavior in a two-choice decision task}
\usage{
plot_twochoiceRL(
  data = NULL,
  subj = 100,
  colors = c("#009999", "#0000FF"),
  plot_type = "static"
)
}
\arguments{
\item{data}{simulated task data}

\item{subj}{subject data to plot. Defaults to subject 100.}

\item{colors}{color to represent the two choices - choice A or choice B.
Defaults to orange for choice A and purple for choice B.}

\item{plot_type}{produce either static or animated version of the plot. Defaults to static.}
}
\value{
A plot of individual \code{subj}'s expected value across trials, a plot of individual \code{subj}'s
  probability across trials, and a side-by-side plot of both of the previous plots.
}
\description{
This function plots the simulated two-choice task data.
}
\examples{

# Save simulated task data to a variable, say, data_sim
data_sim <- simulate_twochoiceRL(trials_unique = FALSE)

# Plot the behavior of individual 100
plot_twochoiceRL(data = data_sim, subj = 100, colors = c("#009999","#0000FF"))
}
