% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_twochoiceRL.R
\name{simulate_twochoiceRL}
\alias{simulate_twochoiceRL}
\title{Simulate behavior in a two-choice decision task}
\usage{
simulate_twochoiceRL(
  seed_value = 528,
  n_subj = 100,
  n_tr = 200,
  x1_mean = 1,
  x1_sd = 1,
  x2_mean = 0,
  x2_sd = 1,
  trials_unique = TRUE
)
}
\arguments{
\item{seed_value}{user-specified random value that stores data information for replication. Defaults to 528.}

\item{n_subj}{number of subjects. Defaults to 100.}

\item{n_tr}{number of trials. Defaults to 200.}

\item{x1_mean}{mean value of the re-parameterized variable for beta. Defaults to 1.}

\item{x1_sd}{standard deviation of the re-parameterized variable for beta. Defaults to 1.}

\item{x2_mean}{mean value of the re-parameterized variable for alpha. Defaults to 0.}

\item{x2_sd}{standard deviation of the re-parameterized variable for beta. Deafults to 1.}

\item{trials_unique}{keep only the unique trials where an individual made a choice. Defaults to TRUE.}
}
\value{
A list containing the two-choice dataset, the number of simulated subjects,
        the 'true' values for parameter 1 (x1) and the 'true' values for parameter 2 (x2). The
        dataset contains the trial number (Trial), the expected value (Value),
        the choice probability (Pr), the presented stimuli (Option), the chosen stimuli (Action),
        and the given reward (Reward).
}
\description{
This function allows users to simulate behavior of \code{n_subj} individuals performing a two-choice task.
}
\examples{
# Simulate 50 subjects performing 100 trials in a two-choice decision task
  data <- simulate_twochoiceRL(n_subj = 50, n_tr = 100)

# View behavioral data of simulated subject 10
  View(data$twochoiceRL[[10]])
}
